Add type check when adding a response to a statement.
